requirements.txt
Flask
flask-cors
flask-jwt-extended
python-dotenv
mysql-connector-python
boto3
werkzeug
requests
